Organizing the tests No more single file chaos and clutter. Test with Arduino still does not work. This new script does not calls Doctests.
