+++ pyfirmata/boards.py
@@ -82,3 +82,4 @@ def pinList2boardDict(pinlist):
     }
 
     return boardDict
+
