+++ test/__init__.py
@@ -1,0 +1,4 @@
+from regression import RegressionTests
+from layout import TestBoardLayout
+from messages import TestBoardMessages
+from boardbases import BoardBaseTest, ArduinoDetection, MockupBoard
